Moved the Coders tests from shelley-ma-test to Test.Data.Coders, added Data.Roundtrip
