+++ eras/alonzo/test-suite/cardano-ledger-alonzo-test.cabal
@@ -95,6 +95,7 @@ test-suite cardano-ledger-alonzo-test
     base16-bytestring,
     bytestring,
     cardano-binary,
+    cardano-data,
     cardano-ledger-alonzo,
     cardano-ledger-alonzo-test,
     cardano-ledger-shelley-ma,
