+++ eras/shelley-ma/test-suite/test/Test/Cardano/Ledger/ShelleyMA/Serialisation.hs
@@ -3,7 +3,6 @@ module Test.Cardano.Ledger.ShelleyMA.Serialisation where
 import Test.Cardano.Ledger.Allegra.Translation (allegraEncodeDecodeTests)
 import Test.Cardano.Ledger.Mary.Translation (maryEncodeDecodeTests)
 import Test.Cardano.Ledger.ShelleyMA.Serialisation.CDDL (cddlTests)
-import Test.Cardano.Ledger.ShelleyMA.Serialisation.Coders (codersTest)
 import Test.Cardano.Ledger.ShelleyMA.Serialisation.Golden.Encoding (goldenEncodingTests)
 import Test.Cardano.Ledger.ShelleyMA.Serialisation.Roundtrip (allEraRoundtripTests)
 import Test.Cardano.Ledger.ShelleyMA.Serialisation.Timelocks (timelockTests)
@@ -14,8 +13,7 @@ tests :: TestTree
 tests =
   testGroup
     "Serialisation tests"
-    [ codersTest,
-      allegraEncodeDecodeTests,
+    [ allegraEncodeDecodeTests,
       maryEncodeDecodeTests,
       txBodyTest,
       timelockTests,
