+++ libs/cardano-data/test/Main.hs
@@ -1,5 +1,6 @@
 module Main where
 
+import Test.Data.Coders (codersTest)
 import Test.Data.UMap (alltests)
 import Test.Tasty
 
@@ -9,7 +10,8 @@ tests :: TestTree
 tests =
   testGroup
     "cardano-data"
-    [ alltests
+    [ alltests,
+      codersTest
     ]
 
 main :: IO ()
